requirements.txt
Flask==3.1.0
peewee==3.17.8
psycopg2==2.9.10
python-dotenv==1.0.1
eclipse-zenoh==1.0.3